% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops_in_or_near_areas.R
\name{area_terms}
\alias{area_terms}
\title{Build the stops_in_or_near_areas table}
\usage{
area_terms(areas_of_interest, area_id_field)
}
\arguments{
\item{areas_of_interest}{}

\item{area_id_field}{}
}
\description{
Build the stops_in_or_near_areas table
}
