% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrivals.R
\name{arrivals}
\alias{arrivals}
\title{Determine the number of arrivals for a single mode at a stop between two time and dates}
\usage{
arrivals(
  gtfs_feed = gtfs_feed,
  stop_id = stop_id,
  start_ymd_hms = start_ymd_hms,
  end_ymd_hms = end_ymd_hms,
  route_type
)
}
\arguments{
\item{gtfs_feed}{A tidygtfs containing the timetable information}

\item{stop_id}{A character string containing the id of the stop}

\item{start_ymd_hms}{A character string describing the start date and time of the period of interest, timezone local to the gtfs files.}

\item{end_ymd_hms}{A character string describing the end date and time of the period of interest, timezone local to the gtfs files.}

\item{route_type}{Numeric describing the type of transportation the number of arrivals is to be calculated for, based on the GTFS static definitions}
}
\description{
Determine the number of arrivals for a single mode at a stop between two time and dates
}
