% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departures.R
\name{departures}
\alias{departures}
\title{Determine the number of departures from a vector of stop ids on a specific date between two times}
\usage{
departures(
  gtfs = gtfs,
  stop_ids = stop_ids,
  date_ymd = date_ymd,
  start_hms = lubridate::hms("-1:0:0"),
  end_hms = lubridate::hms("48:00:00")
)
}
\arguments{
\item{gtfs}{A tidygtfs containing the timetable information.}

\item{stop_ids}{A vector of strings with the stop id numbers from the gtfs.}

\item{date_ymd}{A character string describing the date of interest in yms format.}

\item{start_hms}{A character string describing the start time of the period of interest, timezone local to the gtfs files.}

\item{end_hms}{A character string describing the end time of the period of interest, timezone local to the gtfs files.}
}
\value{
A number.
}
\description{
Determine the number of departures from a vector of stop ids on a specific date between two times
}
