% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly.R
\name{earliest_arrival}
\alias{earliest_arrival}
\title{Returns the time of the earliest arrival in a tidygtfs object}
\usage{
earliest_arrival(gtfs)
}
\arguments{
\item{gtfs}{a tidygtfs}
}
\value{
a lubridate time
}
\description{
Returns the time of the earliest arrival in a tidygtfs object
}
\examples{

#load the revised mornington GTFS data
gtfs = tidytransit::read_gtfs(system.file(
"extdata/mornington180109",
"gtfs.zip", 
package = "gtfssupplyindex", 
mustWork = TRUE))

earliest_arrival(gtfs)


}
