% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departures_and_arrivals.R
\name{first_and_last_stops}
\alias{first_and_last_stops}
\title{Identify first and last stops of each trip}
\usage{
first_and_last_stops(gtfs = gtfs)
}
\arguments{
\item{gtfs}{a tidygtfs object}
}
\value{
a tidygtfs object with additional fields in the trip table for
trip_origin, trip_orgin_id, first_stop_sequence_value, trip_destination,
trip_destination_id and last_stop_sequence_value
}
\description{
Function to identify the first and last stops (id, name and stop_sequence)
for each route in a tidygtfs object
}
