% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nyc_in_metric.R
\name{get_nyc_in_metric}
\alias{get_nyc_in_metric}
\title{Function for creating the nyc example dataset}
\usage{
get_nyc_in_metric()
}
\value{
the nyc boroughs as a sf in metres (crs 32118)
}
\description{
Loads from the nycgeo package and converts to metres. Saves to the inst/extdata directory
}
\examples{
get_nyc_in_metric
}
