% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfssupplyindex-package.R
\docType{package}
\name{gtfssupplyindex-package}
\alias{gtfssupplyindex}
\alias{gtfssupplyindex-package}
\title{gtfssupplyindex: Calculates The Transit Supply Index Based On GTFS and SF Data}
\description{
Tools associated with calculating the Transit Supply Index as per Graham Currie and Zed Senbergs. Identifying spatial gaps in public transport provision for socially disadvantaged Australians: the Melbourne needs-gap study, 2007, Australasian Transport Research Forum (ATRF)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/James-Reynolds/gtfssupplyindex}
  \item Report bugs at \url{https://github.com/James-Reynolds/gtfssupplyindex/issues}
}

}
\author{
\strong{Maintainer}: James Reynolds \email{julianjamesreynolds@gmail.com}

}
\keyword{internal}
