% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly.R
\name{hourly}
\alias{hourly}
\title{Calculate SI for each hour (of service) on a given date, by area_of_interest}
\usage{
hourly(list_gtfs, stops_in_or_near_areas, date_ymd)
}
\arguments{
\item{list_gtfs}{Output of the gtfs_by_route_type function,
a list (by route_type) of tidygtfs objects}

\item{stops_in_or_near_areas}{Output of the stops_in_walk_dist function.
A list (by route_type) of stops and areas within walking distance of each other
and the associated area_terms of SI calculation}

\item{date_ymd}{The date of interest}
}
\value{
A tibble with SI for each hour of service (columns) for each area (rows)
}
\description{
Calculate SI for each hour (of service) on a given date, by area_of_interest
}
\examples{

####-----first load all the inputs
#load the revised mornington GTFS data
list_gtfs = gtfssupplyindex:::gtfs_by_route_type(system.file(
"extdata/mornington180109",
"gtfs.zip", 
package = "gtfssupplyindex", 
mustWork = TRUE))

areas_of_interest <- load_areas_of_interest(absmapsdata::sa22021 \%>\% filter(sa3_name_2021 == "Mornington Peninsula") \%>\% select(sa2_code_2021),  area_id_field = "sa2_code_2021")

buffer_distance <- gtfssupplyindex:::load_buffer_zones()

stops_in_or_near_areas <- gtfssupplyindex:::stops_in_walk_dist(
list_gtfs = list_gtfs, 
areas_of_interest = areas_of_interest,
EPSG_for_transform = 28355,
verbose = FALSE
)

hourly(list_gtfs, stops_in_or_near_areas, "2018-12-30")
}
