% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_areas_of_interest.R
\name{load_areas_of_interest}
\alias{load_areas_of_interest}
\title{Build the areas_of_interest table}
\usage{
load_areas_of_interest(
  areas_of_interest = sf::st_read(system.file("extdata", "nyc.geojson", package =
    "gtfssupplyindex", mustWork = TRUE)),
  area_id_field = "geoid"
)
}
\arguments{
\item{areas_of_interest}{geojson with areas of interest. Defaults to NYC borough boundaries from mfherman/nycgeo package.}

\item{area_id_field}{the primary key field in the areas_of_interest sf. Defaults to geoid, as per NYC borough boundaries}
}
\value{
areas_of_interest table
}
\description{
Build the areas_of_interest table
}
\examples{
load_areas_of_interest()

}
