% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_buffer_zone.R
\name{load_buffer_zones}
\alias{load_buffer_zones}
\title{Loads buffer zones from user defined path or from default csv file}
\usage{
load_buffer_zones(
  buffer_zones_definition_path = system.file("extdata", "buffer_zones.csv", package =
    "gtfssupplyindex", mustWork = TRUE)
)
}
\arguments{
\item{buffer_zones_definition_path}{optional string defining path to a csv defining the buffer zones}
}
\value{
tibble containing route_type, buffer_distance and short_name
}
\description{
Loads buffer zones from user defined path or from default csv file
}
\examples{
load_buffer_zones()

}
