% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI_total.R
\name{si_total}
\alias{si_total}
\title{Total si_by_route_type lit (by route_type) of tables into a si_by_area table}
\usage{
si_total(si_by_route_type)
}
\arguments{
\item{si_by_route_type}{a list (by route_type) of areas and SI scores}
}
\value{
a datatable with areas and total SIs
}
\description{
Having the si by route type may be useful,
but so too is having the total for all modes. This function sums up the SIs
by mode to give a total for each area of interest
}
\examples{
si_total(si_by_route_type)

}
