% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops_in_walk_dist.R
\name{stops_in_walk_dist}
\alias{stops_in_walk_dist}
\title{Create stops_in_or_near_areas table}
\usage{
stops_in_walk_dist(
  list_gtfs,
  areas_of_interest,
  EPSG_for_transform,
  verbose = FALSE
)
}
\arguments{
\item{list_gtfs}{list of tidygtfs (split by route), output of gtfs_by_route_type.R}

\item{areas_of_interest}{sf with areas_of_interest, area_id and geometry}

\item{EPSG_for_transform}{value for the CRS to transform from lat/lon to metres}

\item{verbose}{If TRUE, output of each step will be printed or plotted.}

\item{buffer_distance}{datatable of route_type, buffer_distance and short_name, output of load_buffer_zone.R}
}
\value{
list of dataframes, containing the stops_in_or_near_areas table for each individual route_type
}
\description{
Calculate list of stops_in_or_near_areas tables (by route-type) from  list of tidygtfs (by route_type)
}
\examples{
#load the revised mornington GTFS data
list_gtfs = gtfssupplyindex:::gtfs_by_route_type(system.file(
"extdata/mornington180109",
"gtfs.zip", 
package = "gtfssupplyindex", 
mustWork = TRUE))

areas_of_interest <- load_areas_of_interest(areas_of_interest = sf::st_read(system.file(
"extdata",
"mornington_sa12021.geojson", 
package = "gtfssupplyindex", 
mustWork = TRUE)), 
area_id_field = "sa1_code_2021")

buffer_distance <- gtfssupplyindex:::load_buffer_zones()

gtfssupplyindex:::stops_in_walk_dist(
 list_gtfs = list_gtfs, 
 areas_of_interest = areas_of_interest,
 EPSG_for_transform = 28355, 
 verbose = TRUE)
  
}
