% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_areas_of_interest.R
\name{load_areas_of_interest}
\alias{load_areas_of_interest}
\title{Build the areas_of_interest table}
\usage{
load_areas_of_interest(
  areas_of_interest = sf::st_read(system.file("extdata", "mornington_sa32021.geojson",
    package = "gtfssupplyindex", mustWork = TRUE)),
  area_id_field = "sa3_code_2021"
)
}
\arguments{
\item{areas_of_interest}{sf containing boundaries for which to calculate the SI.
Defaults to load mornington_sa3201 boundaries from extdata.}

\item{area_id_field}{the primary key field in the boundaries sf.
Defaults to sa3_code_2021 as per mornington dataset}
}
\value{
areas_of_interest table as an sf object,
containing only area_id and geometry fields
}
\description{
Build the areas_of_interest table
}
\examples{
load_areas_of_interest()

}
