% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_by_route_type.R, R/load_buffer_zones.R
\name{load_buffer_zones}
\alias{load_buffer_zones}
\title{tidygtfs object to list by route_type of tidygtfs object}
\usage{
load_buffer_zones(
  buffer_zones_definition_path = "inst/extdaya/buffer_zones.csv"
)

load_buffer_zones(
  buffer_zones_definition_path = "inst/extdaya/buffer_zones.csv"
)
}
\arguments{
\item{buffer_zones_definition_path}{String describing the path to the buffer_zone definition file. Defaults to the buffer_zones.csv file included in the packages data directory.}

\item{path}{to GTFS .zip file}
}
\value{
A list, each element being a tidygtfs for an individual mode.

a tibble containing the buffer_distance table,  which has an entry for each route type defining its buffer_distance (typically 400 or 800m) and the short_name of each route type
}
\description{
\code{gtfs_by_route_type()} loads a GTFS into the datatable format used by gtfstool, then splits it by route_type, and returns a list with a tidygtfs for each route type

\code{load_buffer_zones ()} loads information from the buffer_zones.csv about route_types, their names and the associated walking distance thresholds.
}
