% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops_in_walk_dist.R
\name{stops_in_walk_dist}
\alias{stops_in_walk_dist}
\title{Create stops_in_or_near_areas table}
\usage{
stops_in_walk_dist(
  list_gtfs,
  areas_of_interest,
  buffer_distance = gtfssupplyindex:::load_buffer_zones(),
  EPSG_for_transform
)
}
\arguments{
\item{list_gtfs}{list of tidygtfs (split by route), output of gtfs_by_route_type.R}

\item{areas_of_interest}{sf with areas_of_interest, area_id and geometry}

\item{buffer_distance}{datatable of route_type, buffer_distance and short_name, output of load_buffer_zone.R}

\item{EPSG_for_transform}{value for the CRS to transform from lat/lon to metres}
}
\value{
list of dataframes, containing the stops_in_or_near_areas table for each individual route_type
}
\description{
Calculate list of stops_in_or_near_areas tables (by route-type) from  list of tidygtfs (by route_type)
}
