% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops_in_walk_dist.R
\name{stops_in_walk_dist}
\alias{stops_in_walk_dist}
\title{Determine which stops are within walking distance of which areas of intersest}
\usage{
stops_in_walk_dist(
  list_gtfs = gtfssupplyindex:::gtfs_by_route_type(path =
    file.path(system.file("extdata", package = "tidytransit"),
    "google_transit_nyc_subway.zip")),
  areas_of_interest = gtfssupplyindex:::load_areas_of_interest(areas_of_interest =
    sf::st_read(system.file("extdata", "nyc.geojson", package = "gtfssupplyindex",
    mustWork = TRUE)), area_id_field = "geoid")
)
}
\arguments{
\item{list_gtfs}{A list, each element being a tidygtfs for an individual mode.}

\item{areas_of_interest}{sf object with area_id field and geometry only}

\item{buffer_distance}{buffer distance definitions}
}
\value{
a stops_in_or_near_areas tibble consisting of the stop_id, area_id and route_type details.
The area_terms are calculated latedin the area_terms function.
}
\description{
Determine which stops are within walking distance of which areas of intersest
}
\examples{
stops_in_walk_dist()
}
