% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops_in_walk_dist.R
\name{stops_in_walk_dist_one_route}
\alias{stops_in_walk_dist_one_route}
\title{Calculate stops_in_or_near_areas table for a single route_type}
\usage{
stops_in_walk_dist_one_route(
  gtfs_single_route_type = list_gtfs[[1]],
  areas_of_interest,
  EPSG_for_transform,
  verbose = FALSE
)
}
\arguments{
\item{gtfs_single_route_type}{a tidygtfs with only one route_type AND the buffer_distance included as list element}

\item{areas_of_interest}{sf object with area_id field}

\item{EPSG_for_transform}{CRS to transform from lat/lon to metres}
}
\value{
dataframe with stop_id, area_id and area_terms for stop_in_or_near_areas
}
\description{
Calculate stops_in_or_near_areas table for a single route_type
}
