% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops_in_walk_dist.R
\name{stops_in_walk_dist_one_route}
\alias{stops_in_walk_dist_one_route}
\title{Calculate stops_in_or_near_areas table for a single route_type}
\usage{
stops_in_walk_dist_one_route(
  gtfs_single_route_type = list_gtfs[[1]],
  areas_of_interest,
  EPSG_for_transform,
  verbose = FALSE
)
}
\arguments{
\item{gtfs_single_route_type}{a tidygtfs with only one route_type AND the buffer_distance included as list element}

\item{areas_of_interest}{sf object with area_id field}

\item{EPSG_for_transform}{CRS to transform from lat/lon to metres}
}
\value{
dataframe with stop_id, area_id and area_terms for stop_in_or_near_areas
}
\description{
Calculate stops_in_or_near_areas table for a single route_type
}
\examples{
#load the revised mornington GTFS data
list_gtfs = gtfssupplyindex:::gtfs_by_route_type(system.file(
"extdata/mornington180109",
"gtfs.zip", 
package = "gtfssupplyindex", 
mustWork = TRUE))

areas_of_interest <- load_areas_of_interest(areas_of_interest = sf::st_read(system.file(
"extdata",
"mornington_sa12021.geojson", 
package = "gtfssupplyindex", 
mustWork = TRUE)), 
area_id_field = "sa1_code_2021")

buffer_distance <- gtfssupplyindex:::load_buffer_zones()

# add buffer_distance_length to list (by route) of tidy_gtfs
list_gtfs <- mapply(function(x, nm){ 
append(x, (buffer_distance_length = buffer_distance[which(
buffer_distance$short_name \%in\% nm), "buffer_distance"]))}, 
list_gtfs, names(list_gtfs), SIMPLIFY=FALSE)
 
 
 # transform areas_of_interest to CRS in metres
  EPSG_for_transform = 28355
 areas_of_interest <- sf::st_transform(areas_of_interest, EPSG_for_transform)
 
 # calculate Area_area terms
 areas_of_interest$area_area <- sf::st_area(areas_of_interest)
  
gtfssupplyindex:::stops_in_walk_dist_one_route(
 gtfs_single_route_type = list_gtfs[[1]], 
 areas_of_interest = areas_of_interest,
 EPSG_for_transform = 28355, 
 verbose = TRUE)
  


}
